backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
openai
aiofiles
python-dotenv
